requirements.txt
requests==2.31.*
dearpygui==1.11.*
pillow==10.4.*
